refactor(blender): clean up dead code, redundant facades, and improve imports structure
- Remove dead _check_health function and E402 silencing hacks in __init__.py.
- Remove redundant matrix_to_transform dynamic re-export facade from blender_to_core.py and update test imports.
- Replace hardcoded 'visual' and 'collision' string literals with PURPOSE_VISUAL and PURPOSE_COLLISION constants in translator.py.
- Move if TYPE_CHECKING blocks to the top of modules to fully comply with PEP 8 import styling guidelines.
- Clean up unused imports, dead variables, and redundant dynamic imports across adapters and operators.
- Run formatters and test suites to verify 100% green compliance.
