+++ platforms/blender/src/linkforge/blender/handlers/name_sync_handler.py
@@ -8,8 +8,6 @@ from __future__ import annotations
 
 import typing
 
-if typing.TYPE_CHECKING:
-    pass
 from linkforge.core._utils.string_utils import sanitize_name
 
 from ..utils.property_helpers import (
