+++ platforms/blender/src/linkforge/blender/utils/transform_utils.py
@@ -4,6 +4,43 @@ from __future__ import annotations
 
 from typing import Any
 
+from linkforge.core import Transform, Vector3
+from linkforge.core._utils.math_utils import clean_float
+
+try:
+    from mathutils import Matrix
+except ImportError:
+    Matrix = None  # type: ignore[assignment,misc]
+
+
+def matrix_to_transform(matrix: Any) -> Transform:
+    """Convert a Blender 4x4 matrix to a Core Transform.
+
+    Args:
+        matrix: Blender mathutils.Matrix (4x4)
+
+    Returns:
+        Core Transform with XYZ position and RPY rotation.
+    """
+    if matrix is None or Matrix is None:
+        return Transform.identity()
+
+    translation = matrix.to_translation()
+    rotation = matrix.to_euler("XYZ")
+
+    return Transform(
+        xyz=Vector3(
+            clean_float(translation.x),
+            clean_float(translation.y),
+            clean_float(translation.z),
+        ),
+        rpy=Vector3(
+            clean_float(rotation.x),
+            clean_float(rotation.y),
+            clean_float(rotation.z),
+        ),
+    )
+
 
 def set_parent_keep_transform(child_obj: Any, parent_obj: Any) -> None:
     """Set object parent while preserving its world transform (visual location/rotation).
